+++ rust/guest/k16-abi/src/lib.rs
@@ -417,6 +417,7 @@ pub mod computer {
         pub const COMMAND_COPY_TO_USER: i32 = 6;
         pub const COMMAND_SET_TRAP_RETURN_PHYSICAL: i32 = 7;
         pub const COMMAND_SET_TRAP_RETURN_ADDRESS_SPACE: i32 = 8;
+        pub const COMMAND_DESTROY_ADDRESS_SPACE: i32 = 9;
         pub const FLAG_USER_ACCESSIBLE: i32 = 0x0000_0001;
         pub const FLAG_WRITABLE: i32 = 0x0000_0002;
         pub const FLAG_EXECUTABLE: i32 = 0x0000_0004;
@@ -514,6 +515,7 @@ mod tests {
         assert_eq!(computer::mmu0::COMMAND_COPY_TO_USER, 6);
         assert_eq!(computer::mmu0::COMMAND_SET_TRAP_RETURN_PHYSICAL, 7);
         assert_eq!(computer::mmu0::COMMAND_SET_TRAP_RETURN_ADDRESS_SPACE, 8);
+        assert_eq!(computer::mmu0::COMMAND_DESTROY_ADDRESS_SPACE, 9);
         assert_eq!(computer::mmu0::FLAG_USER_ACCESSIBLE, 0x0000_0001);
         assert_eq!(computer::mmu0::FLAG_WRITABLE, 0x0000_0002);
         assert_eq!(computer::mmu0::FLAG_EXECUTABLE, 0x0000_0004);